Hide squash checkbox when there is a single commit
When there is only one commit, the squash checkbox doesn't do
anything. When a commit is added after the widget is loaded, the MR
can't be merged anyway - you have to refresh the page. So we can only
show the checkbox on page load if there's more than one commit.