Remove issue and MR counts from labels index
We were performing two (quite slow, because of the need to do access
checks) queries for each label on the page: one to get the count of
issues with that label, and one to get the count of merge requests with
that label.

Until we have a performant way of doing this, remove the counts.