Use sparse checkout for squashes
For large repos, the checkout as part of creating a worktree can be very
expensive. This can be so expensive that processes get 'stuck' on
GitLab.com.

To make this cheaper, configure sparse checkout for squash worktrees so
that we only check out the files that were changed in the target
branch (new files don't need to be included).

Also add specs to cover the case where only new files are added, and the
case where the diff isn't shown in the merge request because it's too
large.