Remove now unnecessary `additional_rules!` method from GroupPolicy
The intention was for this to be used to override group policy in
EE. This led to all sort of strange problems in practice - for
example, https://gitlab.com/gitlab-org/gitlab-ee/merge_requests/1025
needed to override `owner` permissions, but `additional_rules!`
passed `master` only. The better solution is to use an EE
module to override the policy. Removing this now defunct method.