Fix LDAP DN case-mismatch bug in LDAP group sync
In a few circumstances, we saw that the LDAP person would
present the DN in one mix of case while a given LDAP group
member DN list would return the DN in another mix of case.
LDAP group sync was unable to make the match between these
members and it caused created/updated timestamp to be reset
and LDAP overrides to be ignored.