Add gitaly server dependency
In GitLab 9.0.0 we want to turn on Gitaly in Omnibus installations.
Gitaly is not a required component yet for source installations
because we want to iterate on the configuration and architecture some
more. Omnibus installations will have the option to completely disable
Gitaly if they want to or if necessary.

The new GITALY_SERVER_VERSION file which version of the Gitaly server
application is needed by the current GitLab version.