Fix autocomplete in EE when permissions are sent
In EE, `@users` can be an array at the point where we remove the current
user, because it can do a permissions check on the users for project
mirroring.

Fix this in CE by only using array methods, not AR methods. We can't use
`delete` because that uses `Object#equal?`, which isn't true in this
case.