Send token of depenent job
Artifacts download for authorization is using a job token of job to
which the artifact belongs. In APIv1 the token was sent with dependent
jobs details and in APIv4 it was designed to also contain it.

However I forgot about this parameter while working on
`/api/v4/jobs/request` endpoint.

This commit adds a missing parameter which is required for APIv4 to work
properly.