Improve resiliency of build picking
Every time we pick a new build, or we get into conflict we force runner to retry the operation to make sure that all builds are being picked.

It makes us to loose one request, but also makes sure that we are consistent.

It is fair trade, as number of generated builds is relatively small 0.1% compared to the traffic.