Fix todos page when commit todos are present
We need to preload both:
- The target's project, route, and namespace
- The todo's project, route, and namespace

That's because the view helpers will sometimes need the target's
project, and sometimes need the project - even though they will always
be the same.

Also, don't preload the target for commit todos, as commits are a
special type of target that doesn't exist in the database.