Allow chat notifications only for the default branch
Right now, it once  a chat notifacation service has been enabled, there
is no way to limit the branches which trigger a notification. Instead of
allowing the user to specify the list, I opted to let the user check the
box if they'd only want to be notified of the default branch.

Tags are uneffected by this change.