Do nothing if ci_builds.job_id isn't there
This is really weird. I think we did get the same issue while
migrating on GitLab.com. We fixed this by adding job_id to
ci_builds table manually, and then run the migrations again.
However I think we didn't hit into this on staging, which should
somehow be a production clone.

At any rate, I guess we could check if the column exists in the
migration. If the column is not there, there's no point to remove
that column anyway.

Closes #29976