Speed up suggested approvers
Instead of performing `git log` for each file that's changed, do it for
all files at once. This stops git having to load large packfiles once
for each file in an MR.

This is always a suggestion, so the change in behaviour should be
fine. Increase the number of commits counted by squaring it, to allow
for the fewer calls to `git log`.