Issuable comment & close consistent behaviour
Previously the behaviour of the comment & close action would rely on the
PUT request being slow. If the PUT request was actually faster the
comment may never exist.

This changes that by waiting for the comment to be successfully posted
before sending the PUT request. This behaviour is also mirrored between
both merge requests & issues.

Closes #30577