Add foreign key on trigger requests
I started out as I missed the foreign key on ci_triggers and projects.
However, it seemed Yorick caught this already on gitlab-org/gitlab-ce!9388.

Still I found a missing foreign key, between the trigger requests and
triggers. This commit adds it, and removes the depend: :destroy on the
Rails model Ci::Triggers