Add retry to system hook worker
The default number of Sidekiq retries is 25 over about 21 days.
If an external system is unavailable for a length of time, hooks
can pile up in the Sidekiq queues. If the situation gets bad enough
the retry jobs dropping back into the main queue can slow down
other jobs. Limit the retries to a sane number to avoid this
scenario.