Eliminate N+1 queries in loading namespaces for every issuable in milestones
If we rely on the helper functions in `GitlabRoutingHelper` for `merge_request_path`,
we end up calling a database query to look up the Namespace association for
every merge request since `entity.project.namespace` is called. By reusing the project
defined in the controller, we avoid that problem.

Partial fix to #27387