Migrate data from trigger schedules to pipeline schedules
The models innerworkings and name are changed, so the included
migrations moves the data between the tables if applicable. This
migration is not reversable, as the data needs to be removed from the
trigger table to have a state the end user expects.

The foreigh key will remove the data from the ci_trigger_schedules table
so this table will be empty and can be removed in the next commit.