Fix bug where Geo secondary Sidekiq cron jobs would not be activated if settings changed
If a secondary starts up before it has been made a secondary, the Sidekiq initializer
would disable the backfill and file transfer cron jobs. This state persists indefinitely
until either the entire Redis database is flushed or the jobs are manually activated.
To fix this, we activate the jobs at startup if Geo is being used.

Closes #2289