Fix Import Export tests, reintroduce triggers factory
Because the pipeline failed, both changes were needed to not break the
rest of GitLab.

Pipeline schedules are now import or exported, when using this feature
on your project.

The triggers factory is now back, after it had been removed in a
previous commit on this branch. This broke some tests, and was actually
better than the replacement I introduced.