Convert nested groups to regular ones for MySQL
This migration will take all nested groups and convert them into regular
groups, ensuring that members of any parent groups still have access to
the child group. This migration relies on code external to it as copying
all of this over involves hundreds of lines of code depending on all
sorts of methods, making this practically impossible to do right.