Fallback to default pattern when note does not belong to project
In commit 020295ff a regex was added to quickly determine if mentionables might
be referenced. But this didn't take into account the note might not belong to a
project. So when it doesn't belong to a project, assume the default issue
tracker is used, and use DEFAULT_PATTERN to check if there are mentionables.

Fixes gitlab-org/gitlab-ce#31993