Update DB requirements in the documentation
This makes it more obvious that we greatly prefer PostgreSQL over
MySQL/MariaDB and that all users are recommended to use PostgreSQL
instead. This also removes the paragraph mentioning the size per user
row, as this is completely inaccurate and doesn't really help users with
anticipating the database size. Instead we now simply recommend users to
have at least 5-10 GB of storage at minimum.

[ci skip]