Validate namespace synced with user
A lot of test setups treated User Namespaces as if they could standalone, and have a different `path` and `name` than their owner’s `username`. At some point, that became an unreasonable assumption.

This new validation reflects the new world and enforces a matching `path`, `name`, and `username`, so the tests must change. In many cases, it is easier to change the test to use a Group Namespace since the `path` and `name` can be set at-will.