Override `file_storage` on the `ObjectStoreUploader`
Normally the storage-stype of an uploader is set on the class level of
an uploader [0] and the `storage` property is initialized from there [1].

In a security release, we were using that at the class level to
determine the path of an upload and whether or not to move it.

In the `ObjectStoreUploader` the storage method is overridden, meaning
that the storage-type set on the class could be different from the
actual storage used.

[0] https://github.com/carrierwaveuploader/carrierwave/blob/f2f8569b1013f601fe7be4b3a616359acd27240d/lib/carrierwave/uploader/configuration.rb#L8
[1] https://github.com/carrierwaveuploader/carrierwave/blob/f2f8569b1013f601fe7be4b3a616359acd27240d/lib/carrierwave/uploader/store.rb#L88