Add in_review_folder to usage ping
As its hard to reliably check how many review apps there are on the
clients machine, we start by checking where the type is `review`. This
means the folder is called that way. This will lead to a seq
scan on the table. However, this is done once a week, so the benefit of
adding an index seems not to apply here.