Speed up counting approvers when some are specified
When some approvers are specified, we still need to check the potential set of
approvers, including people with developer access and up. Doing that with an OR
took a long time (over 500ms) on GitLab.com, even when only one approver was
specified.

By doing a UNION instead, we get much faster results (less than 10ms). Also, as
we are using UNION and not UNION ALL, we don't need to exclude the approvers
specified from the query to get everyone with access.