Allow API access with CI_JOB_TOKEN
In the current implementation the `initial_current_user` can be set to
the user whom triggered the job if the token matches the CI_JOB_TOKEN.
This however grants the key to kingdom, which also might be a feature.
To limit the exposure and make interation less likely, the token is
NULLIFIED when the job is finished.

The attack vector now does include all jobs which either
- Playable jobs
- Jobs which never finish somehow